refactor: update blog
